+++ packages/catalog-client/src/CatalogClient.ts
@@ -39,9 +39,13 @@ import {
   ValidateEntityResponse,
   GetEntitiesByRefsRequest,
   GetEntitiesByRefsResponse,
+  QueryEntitiesRequest,
+  QueryEntitiesResponse,
+  EntityFilterQuery,
 } from './types/api';
 import { DiscoveryApi } from './types/discovery';
 import { FetchApi } from './types/fetch';
+import { isQueryEntitiesInitialRequest } from './utils';
 
 /**
  * A frontend and backend compatible client for communicating with the Backstage
@@ -107,30 +111,7 @@ export class CatalogClient implements CatalogApi {
       limit,
       after,
     } = request ?? {};
-    const params: string[] = [];
-
-    // filter param can occur multiple times, for example
-    // /api/catalog/entities?filter=metadata.name=wayback-search,kind=component&filter=metadata.name=www-artist,kind=component'
-    // the "outer array" defined by `filter` occurrences corresponds to "anyOf" filters
-    // the "inner array" defined within a `filter` param corresponds to "allOf" filters
-    for (const filterItem of [filter].flat()) {
-      const filterParts: string[] = [];
-      for (const [key, value] of Object.entries(filterItem)) {
-        for (const v of [value].flat()) {
-          if (v === CATALOG_FILTER_EXISTS) {
-            filterParts.push(encodeURIComponent(key));
-          } else if (typeof v === 'string') {
-            filterParts.push(
-              `${encodeURIComponent(key)}=${encodeURIComponent(v)}`,
-            );
-          }
-        }
-      }
-
-      if (filterParts.length) {
-        params.push(`filter=${filterParts.join(',')}`);
-      }
-    }
+    const params = this.getParams(filter);
 
     if (fields.length) {
       params.push(`fields=${fields.map(encodeURIComponent).join(',')}`);
@@ -224,6 +205,64 @@ export class CatalogClient implements CatalogApi {
   }
 
   /**
+   * {@inheritdoc CatalogApi.queryEntities}
+   */
+  async queryEntities(
+    request: QueryEntitiesRequest = {},
+    options?: CatalogRequestOptions,
+  ) {
+    const params: string[] = [];
+
+    if (isQueryEntitiesInitialRequest(request)) {
+      const {
+        fields = [],
+        filter,
+        limit,
+        orderFields,
+        fullTextFilter,
+      } = request;
+      params.push(...this.getParams(filter));
+
+      if (limit !== undefined) {
+        params.push(`limit=${limit}`);
+      }
+      if (orderFields !== undefined) {
+        (Array.isArray(orderFields) ? orderFields : [orderFields]).forEach(
+          ({ field, order }) => params.push(`sortField=${field},${order}`),
+        );
+      }
+      if (fields.length) {
+        params.push(`fields=${fields.map(encodeURIComponent).join(',')}`);
+      }
+
+      const normalizedFullTextFilterTerm = fullTextFilter?.term?.trim();
+      if (normalizedFullTextFilterTerm) {
+        params.push(`fullTextFilterTerm=${normalizedFullTextFilterTerm}`);
+      }
+      if (fullTextFilter?.fields?.length) {
+        params.push(`fullTextFilterFields=${fullTextFilter.fields.join(',')}`);
+      }
+    } else {
+      const { fields = [], limit, cursor } = request;
+
+      params.push(`cursor=${cursor}`);
+      if (limit !== undefined) {
+        params.push(`limit=${limit}`);
+      }
+      if (fields.length) {
+        params.push(`fields=${fields.map(encodeURIComponent).join(',')}`);
+      }
+    }
+
+    const query = params.length ? `?${params.join('&')}` : '';
+    return this.requestRequired<QueryEntitiesResponse>(
+      'GET',
+      `/entities/by-query${query}`,
+      options,
+    );
+  }
+
+  /**
    * {@inheritdoc CatalogApi.getEntityByRef}
    */
   async getEntityByRef(
@@ -290,30 +329,7 @@ export class CatalogClient implements CatalogApi {
     options?: CatalogRequestOptions,
   ): Promise<GetEntityFacetsResponse> {
     const { filter = [], facets } = request;
-    const params: string[] = [];
-
-    // filter param can occur multiple times, for example
-    // /api/catalog/entities?filter=metadata.name=wayback-search,kind=component&filter=metadata.name=www-artist,kind=component'
-    // the "outer array" defined by `filter` occurrences corresponds to "anyOf" filters
-    // the "inner array" defined within a `filter` param corresponds to "allOf" filters
-    for (const filterItem of [filter].flat()) {
-      const filterParts: string[] = [];
-      for (const [key, value] of Object.entries(filterItem)) {
-        for (const v of [value].flat()) {
-          if (v === CATALOG_FILTER_EXISTS) {
-            filterParts.push(encodeURIComponent(key));
-          } else if (typeof v === 'string') {
-            filterParts.push(
-              `${encodeURIComponent(key)}=${encodeURIComponent(v)}`,
-            );
-          }
-        }
-      }
-
-      if (filterParts.length) {
-        params.push(`filter=${filterParts.join(',')}`);
-      }
-    }
+    const params = this.getParams(filter);
 
     for (const facet of facets) {
       params.push(`facet=${encodeURIComponent(facet)}`);
@@ -466,11 +482,11 @@ export class CatalogClient implements CatalogApi {
     }
   }
 
-  private async requestRequired(
+  private async requestRequired<T = any>(
     method: string,
     path: string,
     options?: CatalogRequestOptions,
-  ): Promise<any> {
+  ): Promise<T> {
     const url = `${await this.discoveryApi.getBaseUrl('catalog')}${path}`;
     const headers: Record<string, string> = options?.token
       ? { Authorization: `Bearer ${options.token}` }
@@ -481,7 +497,7 @@ export class CatalogClient implements CatalogApi {
       throw await ResponseError.fromResponse(response);
     }
 
-    return await response.json();
+    return response.json();
   }
 
   private async requestOptional(
@@ -503,5 +519,32 @@ export class CatalogClient implements CatalogApi {
     }
 
     return await response.json();
+  }
+
+  private getParams(filter: EntityFilterQuery = []) {
+    const params: string[] = [];
+    // filter param can occur multiple times, for example
+    // /api/catalog/entities?filter=metadata.name=wayback-search,kind=component&filter=metadata.name=www-artist,kind=component'
+    // the "outer array" defined by `filter` occurrences corresponds to "anyOf" filters
+    // the "inner array" defined within a `filter` param corresponds to "allOf" filters
+    for (const filterItem of [filter].flat()) {
+      const filterParts: string[] = [];
+      for (const [key, value] of Object.entries(filterItem)) {
+        for (const v of [value].flat()) {
+          if (v === CATALOG_FILTER_EXISTS) {
+            filterParts.push(encodeURIComponent(key));
+          } else if (typeof v === 'string') {
+            filterParts.push(
+              `${encodeURIComponent(key)}=${encodeURIComponent(v)}`,
+            );
+          }
+        }
+      }
+
+      if (filterParts.length) {
+        params.push(`filter=${filterParts.join(',')}`);
+      }
+    }
+    return params;
   }
 }