+++ plugins/catalog-common/src/index.ts
@@ -21,19 +21,6 @@
  * @packageDocumentation
  */
 
-export {
-  RESOURCE_TYPE_CATALOG_ENTITY,
-  catalogEntityReadPermission,
-  catalogEntityCreatePermission,
-  catalogEntityDeletePermission,
-  catalogEntityRefreshPermission,
-  catalogLocationReadPermission,
-  catalogLocationCreatePermission,
-  catalogLocationDeletePermission,
-  catalogPermissions,
-} from './permissions';
-export type { CatalogEntityPermission } from './permissions';
-
 export * from './search';
 export * from './ingestion';
 export type { LocationSpec } from './common';