+++ plugins/scaffolder-backend/src/modules/catalogModuleTemplateKind.test.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import { catalogProcessingExtensionPoint } from '@backstage/plugin-catalog-node';
+import { catalogProcessingExtensionPoint } from '@backstage/plugin-catalog-node/alpha';
 import { ScaffolderEntitiesProcessor } from '../processor';
 import { catalogModuleTemplateKind } from './catalogModuleTemplateKind';
 import { startTestBackend } from '@backstage/backend-test-utils';