+++ packages/backend-defaults/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @backstage/backend-defaults
 
+## 0.1.8-next.0
+
+### Patch Changes
+
+- 928a12a9b3: Internal refactor of `/alpha` exports.
+- Updated dependencies
+  - @backstage/backend-plugin-api@0.4.1-next.0
+  - @backstage/backend-app-api@0.4.1-next.0
+
 ## 0.1.7
 
 ### Patch Changes