+++ packages/backend-plugin-api/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @backstage/backend-plugin-api
 
+## 0.4.1-next.0
+
+### Patch Changes
+
+- 928a12a9b3: Internal refactor of `/alpha` exports.
+- Updated dependencies
+  - @backstage/backend-tasks@0.4.4-next.0
+  - @backstage/config@1.0.6
+  - @backstage/types@1.0.2
+  - @backstage/plugin-auth-node@0.2.12-next.0
+  - @backstage/plugin-permission-common@0.7.3
+
 ## 0.4.0
 
 ### Minor Changes