+++ packages/dev-utils/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @backstage/dev-utils
 
+## 1.0.13-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-catalog-react@1.4.0-next.0
+  - @backstage/core-plugin-api@1.4.1-next.0
+  - @backstage/catalog-model@1.2.1-next.0
+  - @backstage/test-utils@1.2.6-next.0
+  - @backstage/app-defaults@1.2.1-next.0
+  - @backstage/core-app-api@1.5.1-next.0
+  - @backstage/core-components@0.12.5-next.0
+  - @backstage/integration-react@1.1.11-next.0
+  - @backstage/theme@0.2.17
+
 ## 1.0.12
 
 ### Patch Changes