+++ packages/techdocs-cli-embedded-app/CHANGELOG.md
@@ -1,5 +1,24 @@
 # techdocs-cli-embedded-app
 
+## 0.2.80-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/cli@0.22.4-next.0
+  - @backstage/plugin-techdocs@1.6.0-next.0
+  - @backstage/core-plugin-api@1.4.1-next.0
+  - @backstage/catalog-model@1.2.1-next.0
+  - @backstage/plugin-techdocs-react@1.1.4-next.0
+  - @backstage/test-utils@1.2.6-next.0
+  - @backstage/plugin-catalog@1.9.0-next.0
+  - @backstage/app-defaults@1.2.1-next.0
+  - @backstage/config@1.0.6
+  - @backstage/core-app-api@1.5.1-next.0
+  - @backstage/core-components@0.12.5-next.0
+  - @backstage/integration-react@1.1.11-next.0
+  - @backstage/theme@0.2.17
+
 ## 0.2.79
 
 ### Patch Changes