+++ plugins/auth-node/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @backstage/plugin-auth-node
 
+## 0.2.12-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-common@0.18.3-next.0
+  - @backstage/config@1.0.6
+  - @backstage/errors@1.1.4
+
 ## 0.2.11
 
 ### Patch Changes