+++ plugins/azure-sites/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @backstage/plugin-azure-sites
 
+## 0.1.5-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-catalog-react@1.4.0-next.0
+  - @backstage/core-plugin-api@1.4.1-next.0
+  - @backstage/catalog-model@1.2.1-next.0
+  - @backstage/core-components@0.12.5-next.0
+  - @backstage/theme@0.2.17
+  - @backstage/plugin-azure-sites-common@0.1.0
+
 ## 0.1.4
 
 ### Patch Changes