+++ plugins/catalog-backend-module-msgraph/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @backstage/plugin-catalog-backend-module-msgraph
 
+## 0.5.2-next.0
+
+### Patch Changes
+
+- a4ebd51352: Revert using `baseUrl` for MS Graph provider as it breaks token retrieval
+- 928a12a9b3: Internal refactor of `/alpha` exports.
+- 2c234a8932: properly support custom graph api URL
+- Updated dependencies
+  - @backstage/plugin-catalog-backend@1.8.0-next.0
+  - @backstage/backend-tasks@0.4.4-next.0
+  - @backstage/backend-plugin-api@0.4.1-next.0
+  - @backstage/backend-common@0.18.3-next.0
+  - @backstage/catalog-model@1.2.1-next.0
+  - @backstage/plugin-catalog-node@1.3.4-next.0
+  - @backstage/config@1.0.6
+
 ## 0.5.0
 
 ### Minor Changes