+++ plugins/entity-feedback-backend/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @backstage/plugin-entity-feedback-backend
 
+## 0.1.1-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/catalog-client@1.4.0-next.0
+  - @backstage/backend-common@0.18.3-next.0
+  - @backstage/catalog-model@1.2.1-next.0
+  - @backstage/config@1.0.6
+  - @backstage/plugin-auth-node@0.2.12-next.0
+  - @backstage/plugin-entity-feedback-common@0.1.0
+
 ## 0.1.0
 
 ### Minor Changes