+++ plugins/firehydrant/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @backstage/plugin-firehydrant
 
+## 0.1.33-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-catalog-react@1.4.0-next.0
+  - @backstage/core-plugin-api@1.4.1-next.0
+  - @backstage/core-components@0.12.5-next.0
+  - @backstage/theme@0.2.17
+
 ## 0.1.32
 
 ### Patch Changes