+++ plugins/permission-backend/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @backstage/plugin-permission-backend
 
+## 0.5.18-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-common@0.18.3-next.0
+  - @backstage/config@1.0.6
+  - @backstage/errors@1.1.4
+  - @backstage/plugin-auth-node@0.2.12-next.0
+  - @backstage/plugin-permission-common@0.7.3
+  - @backstage/plugin-permission-node@0.7.6-next.0
+
 ## 0.5.17
 
 ### Patch Changes