+++ plugins/todo-backend/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @backstage/plugin-todo-backend
 
+## 0.1.40-next.0
+
+### Patch Changes
+
+- 4120513412: todo-backend is now exposed as a plugin which uses the new plugin system
+- Updated dependencies
+  - @backstage/catalog-client@1.4.0-next.0
+  - @backstage/backend-plugin-api@0.4.1-next.0
+  - @backstage/backend-common@0.18.3-next.0
+  - @backstage/catalog-model@1.2.1-next.0
+  - @backstage/plugin-catalog-node@1.3.4-next.0
+  - @backstage/config@1.0.6
+  - @backstage/errors@1.1.4
+  - @backstage/integration@1.4.2
+
 ## 0.1.39
 
 ### Patch Changes