+++ plugins/user-settings-backend/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @backstage/plugin-user-settings-backend
 
+## 0.1.7-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-plugin-api@0.4.1-next.0
+  - @backstage/backend-common@0.18.3-next.0
+  - @backstage/catalog-model@1.2.1-next.0
+  - @backstage/errors@1.1.4
+  - @backstage/types@1.0.2
+  - @backstage/plugin-auth-node@0.2.12-next.0
+
 ## 0.1.6
 
 ### Patch Changes